packages/toolpacks/pyproject.toml
[project]
name = "nai-toolpacks"
version = "0.1.0"
description = "Specialized content extractors for nAI: PDF, Web, Email, Code"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "AGPL-3.0-only"}
authors = [
    {name = "Nithron Team", email = "dev@nithron.io"}
]
keywords = ["extraction", "pdf", "ocr", "web-scraping", "email", "code-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = []

[project.optional-dependencies]
pdf = [
    "pypdf>=4.0",
    "pdf2image>=1.16",
    "pytesseract>=0.3",
    "pillow>=10.0",
]
web = [
    "aiohttp>=3.9",
    "requests>=2.31",
    "playwright>=1.40",
]
email = [
    # Standard library is sufficient
]
code = [
    # Standard library is sufficient
]
all = [
    "pypdf>=4.0",
    "pdf2image>=1.16",
    "pytesseract>=0.3",
    "pillow>=10.0",
    "aiohttp>=3.9",
    "requests>=2.31",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/Nithronverse/nAI"
Documentation = "https://github.com/Nithronverse/nAI/tree/main/packages/toolpacks"
Repository = "https://github.com/Nithronverse/nAI"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

